Enable cross-seed deletion/retention
Cross-seeds are detected by torrents that share the same root path

Cross-seed handling is enabled if the config `delete_data` is set to
`True`. If no data is deleted the cross-seed is handled by another
strategy. If `delete_data` is set to `True` torrents seeding the same
data as a torrent marked for deletion will fail.

If enabled, deletion is by defailt not propagated to the cross-seeded
torrents, instead the torrent marked for removal is flagged as
soft-remove, meaning that the torrent will be deleted but not the
data.

By changing the flag `remove_cross_seeds` in the strategy
configuration to `all_trackers` or a list of tracker host names
deletion will propagate to either all or the selected
trackers. Torrents included in the proagation will also be deleted,
even if not included by the filter.

Support has been added to:
 * deluge
 * qBittorrent APIv2
 * Transmission

A new field has been added to `Torrent#base_path`