change: remove Outgoing Port Status checker
We have confirmed a bug, which is, the Outgoing Port Status checker will
fail and report 'portTested: http error 400: Bad Request' when we are
using Transmission and check the outgoing port status in IPv6 Network.

This bug will make the task failed, however, there are no removing
conditions relying on this port status. Therefore we remove it.

Closes #101, #135.

This bug is in Transmission, which can be referred to
transmission/transmission#3102, not in our software.